Fixed function parameter type

+++ TrafficModel/binding.cc
@@ -56,15 +56,15 @@ PYBIND11_MODULE(TrafficModel,m) {
 
 	// VehicleType class
 	py::class_<VehicleType>(m,"VehicleType")
-		.def(py::init<int,int>())
+		.def(py::init<uint16_t,uint16_t>())
 		.def("getSource", &VehicleType::getSource)
 		.def("getDestination", &VehicleType::getDestination)
 		.def("setTransMatrix", &VehicleType::setTransMatrix);
 		
 	// Vehicle class
 	py::class_<Vehicle>(m,"Vehicle")
-		.def(py::init<int>())
-		.def_static("addVehicleType", static_cast<void (*)(uint8_t,uint8_t)>(&Vehicle::addVehicleType))
+		.def(py::init<uint8_t>())
+		.def_static("addVehicleType", static_cast<void (*)(uint16_t,uint16_t)>(&Vehicle::addVehicleType))
 		.def_static("addVehicleType", static_cast<void (*)(const char *)>(&Vehicle::addVehicleType))
 		.def_static("getVehicleType", &Vehicle::getVehicleType)
 		.def("getType", &Vehicle::getType)
